feat: add outcome evaluation script
