--- scripts/score_history.py
@@ -1,35 +1,0 @@
-#!/usr/bin/env python3
-# scripts/score_history.py
-# - Reads data/history/outcomes.csv
-# - For Outcome == PENDING, checks if TargetLevel was hit (High >= level)
-#   between EvalDate (inclusive) and min(WindowEnd, today) (inclusive).
-#   Results are written back to outcomes.csv.
-import sys
-from pathlib import Path
-
-ROOT = Path(__file__).resolve().parents[1]
-if str(ROOT) not in sys.path:
-    sys.path.insert(0, str(ROOT))
-
-from utils.io import OUTCOMES_CSV
-from utils.outcomes import evaluate_outcomes, read_outcomes, write_outcomes
-
-
-def main() -> None:
-    if not OUTCOMES_CSV.exists():
-        print("No outcomes.csv yet; nothing to score.")
-        return
-
-    df = read_outcomes(OUTCOMES_CSV)
-    if df.empty:
-        print("outcomes.csv empty; nothing to score.")
-        return
-
-    new_df = evaluate_outcomes(df, mode="historical")
-    write_outcomes(new_df, OUTCOMES_CSV)
-    print(f"Scored {len(new_df)} rows → wrote outcomes.csv")
-
-
-if __name__ == "__main__":
-    main()
-
